refactor: extract action panel into primitives
