+++ packages/backend/src/routes/api/v1/history.ts
@@ -101,15 +101,15 @@ historyRouter.post<CanvasIdParam>(
         : undefined;
 
       const pixelHistory = await getPixelHistorySummary(
-      {
+        {
           canvasId,
           points,
           dateRange,
           userIdFilter,
           colorFilter,
         },
-      true,
-    );
+        true,
+      );
 
       res.status(200).json(pixelHistory);
     } catch (error) {
@@ -127,51 +127,51 @@ historyRouter.delete<CanvasIdParam>(
 
       const canvasId = await parseCanvasId(req.params);
 
-    const bodyResult = await PixelHistoryDeleteBodyModel.safeParseAsync(
-      req.body,
-    );
-    assertZodSuccess(bodyResult);
+      const bodyResult = await PixelHistoryDeleteBodyModel.safeParseAsync(
+        req.body,
+      );
+      assertZodSuccess(bodyResult);
 
-    const {
-      x0,
-      y0,
-      x1,
-      y1,
-      fromDateTime,
-      toDateTime,
-      includeUserIds,
-      excludeUserIds,
-      includeColors,
-      excludeColors,
-      shouldBlockAuthors,
-    } = bodyResult.data;
+      const {
+        x0,
+        y0,
+        x1,
+        y1,
+        fromDateTime,
+        toDateTime,
+        includeUserIds,
+        excludeUserIds,
+        includeColors,
+        excludeColors,
+        shouldBlockAuthors,
+      } = bodyResult.data;
 
-    const point0 = { x: x0, y: y0 };
-    const point1 = { x: x1 ?? x0, y: y1 ?? y0 };
+      const point0 = { x: x0, y: y0 };
+      const point1 = { x: x1 ?? x0, y: y1 ?? y0 };
 
-    const userIdFilter =
-      includeUserIds ? { ids: includeUserIds.map(BigInt), include: true }
-      : excludeUserIds ? { ids: excludeUserIds.map(BigInt), include: false }
-      : undefined;
+      const userIdFilter =
+        includeUserIds ? { ids: includeUserIds.map(BigInt), include: true }
+        : excludeUserIds ? { ids: excludeUserIds.map(BigInt), include: false }
+        : undefined;
 
-    const colorFilter =
-      includeColors ? { colors: includeColors, include: true }
-      : excludeColors ? { colors: excludeColors, include: false }
-      : undefined;
+      const colorFilter =
+        includeColors ? { colors: includeColors, include: true }
+        : excludeColors ? { colors: excludeColors, include: false }
+        : undefined;
 
-    await deletePixelHistoryEntries(
-      {
-        canvasId,
-        points: [point0, point1],
-        dateRange: {
-          from: fromDateTime,
-          to: toDateTime,
+      await deletePixelHistoryEntries(
+        {
+          canvasId,
+          points: [point0, point1],
+          dateRange: {
+            from: fromDateTime,
+            to: toDateTime,
+          },
+          userIdFilter,
+          colorFilter,
         },
-        userIdFilter,
-        colorFilter,
-      },
-      shouldBlockAuthors,
-    );
+        shouldBlockAuthors,
+      );
 
       res.status(204).send();
     } catch (error) {
