+++ packages/frontend/src/components/action-panel/primitives.tsx
@@ -1,17 +1,6 @@
-"use client";
-
 import { styled } from "@mui/material";
 import type React from "react";
-import { useId } from "react";
-import {
-  useActionPanelContext,
-  useCanvasContext,
-  useCanvasViewContext,
-  useSelectedColorContext,
-} from "@/contexts";
 import { CANVAS_WRAPPER_CLASS_NAME } from "@/util";
-import { PixelInfoTab, PlacePixelTab } from "./tabs";
-import FramesTab from "./tabs/FramesTab";
 
 export const ActionPanelWrapper = styled("div")`
   --padding-width: 1rem;
@@ -43,14 +32,14 @@ export const ActionPanelWrapper = styled("div")`
   }
 `;
 
-export const TabBar = styled("div")`
+export const ActionPanelTabBar = styled("div")`
   border-radius: 0.5rem;
   display: grid;
   gap: 0.5rem;
   grid-template-columns: repeat(3, 1fr);
 `;
 
-export const StyledTab = styled("button")`
+export const ActionPanelTab = styled("button")`
   appearance: none;
   border: none;
   color: inherit;
@@ -107,7 +96,7 @@ export const StyledTab = styled("button")`
   }
 `;
 
-export const Heading = styled("h2")`
+export const PanelSectionHeading = styled("h2")`
   color: oklch(from var(--discord-white) l c h / 60%);
   font-weight: 600;
   font-size: 1rem;
@@ -117,10 +106,6 @@ export const Heading = styled("h2")`
   text-transform: uppercase;
 `;
 
-type TabKey = "look" | "place" | "frame";
-
-const Tab = GenericTab<TabKey>;
-
 export function GenericTab<TabKey extends string>({
   tabKey,
   onSwitchTab,
@@ -128,9 +113,9 @@ export function GenericTab<TabKey extends string>({
 }: {
   tabKey: TabKey;
   onSwitchTab: (tabKey: TabKey) => void;
-} & React.ComponentPropsWithRef<typeof StyledTab>) {
+} & React.ComponentPropsWithRef<typeof ActionPanelTab>) {
   return (
-    <StyledTab
+    <ActionPanelTab
       onClick={() => onSwitchTab(tabKey)}
       onKeyUp={(event) => {
         if (event.key === "Enter" || event.key === " ") onSwitchTab(tabKey);
@@ -138,90 +123,5 @@ export function GenericTab<TabKey extends string>({
       {...props}
       role="tab"
     />
-  );
-}
-
-export default function ActionPanel() {
-  const {
-    areTabsLocked,
-    currentTab,
-    setAreTabsLocked,
-    setCurrentTab,
-    setTempColor,
-    tempColor,
-  } = useActionPanelContext();
-
-  const { color, setColor } = useSelectedColorContext();
-  const { canvas } = useCanvasContext();
-  const { setIsReticleVisible } = useCanvasViewContext();
-
-  const onSwitchTab = (newTab: TabKey) => {
-    if (areTabsLocked) return;
-
-    setCurrentTab(newTab);
-
-    // hiding colour from reticle if we are on look tab
-    if (newTab === "look") {
-      setTempColor(color);
-      setColor(null);
-    } else {
-      setColor(tempColor);
-    }
-
-    // hiding reticle if we are on frames tab
-    setIsReticleVisible(newTab !== "frame");
-  };
-
-  const placeTabId = useId();
-  const lookTabId = useId();
-  const frameTabId = useId();
-
-  return (
-    <ActionPanelWrapper>
-      <TabBar role="tablist">
-        <Tab
-          aria-controls={placeTabId}
-          aria-disabled={areTabsLocked && currentTab !== "place"}
-          aria-selected={currentTab === "place"}
-          tabKey="place"
-          onSwitchTab={onSwitchTab}
-        >
-          Place
-        </Tab>
-        <Tab
-          aria-controls={lookTabId}
-          aria-disabled={areTabsLocked && currentTab !== "look"}
-          aria-selected={currentTab === "look"}
-          tabKey="look"
-          onSwitchTab={onSwitchTab}
-        >
-          Look
-        </Tab>
-        <Tab
-          aria-controls={frameTabId}
-          aria-disabled={areTabsLocked && currentTab !== "frame"}
-          aria-selected={currentTab === "frame"}
-          tabKey="frame"
-          onSwitchTab={onSwitchTab}
-        >
-          Frame
-        </Tab>
-      </TabBar>
-      <PlacePixelTab
-        active={currentTab === "place"}
-        eventId={canvas.eventId}
-        id={placeTabId}
-      />
-      <PixelInfoTab
-        active={currentTab === "look"}
-        canvasId={canvas.id}
-        id={lookTabId}
-      />
-      <FramesTab
-        active={currentTab === "frame"}
-        id={frameTabId}
-        setTabsLocked={setAreTabsLocked}
-      />
-    </ActionPanelWrapper>
   );
 }
