+++ packages/frontend/src/components/action-panel/tabs/PlacePixelTab.tsx
@@ -11,7 +11,7 @@ import { usePalette, usePlaySound } from "@/hooks";
 import { getUserGuildIds } from "@/util";
 import { DynamicAnchorButton, PlacePixelButton } from "../../button";
 import { InteractiveSwatch } from "../../swatch";
-import { Heading } from "../ActionPanel";
+import { PanelSectionHeading } from "../primitives";
 import {
   ActionPanelTabBody,
   FullWidthScrollView,
@@ -182,7 +182,7 @@ function NamedPalette({ colors, name }: NamedPaletteProps) {
   const isLoading = colors === undefined;
   return (
     <>
-      <Heading>{name}</Heading>
+      <PanelSectionHeading>{name}</PanelSectionHeading>
       <Fieldset>
         {isLoading ?
           Array.from({ length: 12 }).map((_, i) => (
